backend/requirements.txt
Django>=5.2.12
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.10
whitenoise>=6.9.0
